backend/pyproject.toml
[tool.poetry]
name = "findCoach"
version = "1.0"
description = "A Django API for the final project from Vue -- The Complete Guide by Maximilian Schwarzmüller."
authors = ["Christian Gonzalez <christiangonzalezblack@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.12.2"
coreapi = "^2.3.3"
PyYAML = "^5.4.1"
drf-yasg = "^1.20.0"
dj-rest-auth = "^2.1.3"
django-allauth = "^0.44.0"
django-cors-headers = "^3.7.0"
djangorestframework-simplejwt = "^4.6.0"
Django = "^3.1.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
